// src/bot-state-types.ts
// Auto-generated by state:sync - DO NOT EDIT

export type AppStates = 'idle' | 'menu' | 'welcome';

export type StateTransitions = {
  idle: 'welcome' | void;
  menu: 'idle' | 'welcome' | void;
  welcome: 'menu' | void;
};

export type IdleTransitions = Promise<StateTransitions['idle']>;
export type MenuTransitions = Promise<StateTransitions['menu']>;
export type WelcomeTransitions = Promise<StateTransitions['welcome']>;